requirements.txt
fastapi
uvicorn
websockets
black
pylint
aiohttp
aiodns
PyYAML
python-jose[cryptography]
python-multipart
cryptography
passlib[bcrypt]
sqlalchemy
PyMySQL[ed25519]
